Create readme and first solution for Project Euler Problem 2
Created new files for Problem 2 of Project Euler. This problem involves summing the even terms of the Fibonacci Sequence up to four million. A simple HTML page with JavaScript was used. The top of the page includes the verbatim problem description from the Project Euler website and the next section down displays and explains my code, including the solution to the problem.