Determine prime factors and add to array
The current state of the project is able to take a dividend and make an array of its prime factors, but the completeness check needs work as it doesn't recognize when a prime factor is the largest for a given divident. The while loop in the cycleDivisors function is commented out while this gets fixed.