Determine prime factors of arbitrarily large numbers.
This commit meets the criteria laid out by the problem and is able to find the largest prime factor of 600851475143 (which is 6857), along with any other number set as the variable 'dividend'. However, it is slow, taking about 3 minutes to compute the above prime factor. The checkCompleteness function is clumsily written, simply checking if the current divisor being checked is greater than or equal to the dividend divided by the smallest prime factor, and terminating the search if it is. The plan is to replace the current checkCompleteness function with another one that simply divides the dividend by all the elements of primeFactorArray repeatedly until it either arrives at a quotient of 1 or none of elements cleanly divide the running total anymore.