Improve efficiency of checkCompleteness
checkCompleteness now takes dividend and repeatedly divides it by all elements of primeFactorArray until either 1) the result is equal to zero and therefore all prime factors have been found, or 2) the result can no long be cleanly divided by any of the prime factors and therefore one or more prime factors remains. This is much faster than the previous solution, which took 2 minutes 33 seconds to find the largest prime of 600851475143. This solution returned the result instantly.