Calculate largest palindrome product of two three-digit numbers
This program rapidly computes the largest possible product of two three-digit numbers that is also a palindrome (reads the same forward and backward.

This solution starts by initiating two for loops, one for the first three-digit number (firstNum), counting down from 999, and the other for the second three-digit number(secondNum), counting down from the current value of the first number. Both loops are set to terminate when the number reaches a floor (floor) that is initially set at 100 (the lowest possible three-digit number).

It then multiplies the numbers together and checks if the product is a palindrome. If so, it checks to see if the product is larger than the current largest palindrome found (bigPal). If so, it sets the product as the new value of bigPal and sets the current secondNum as the new floor.

This solution works quickly by 1) immediately eliminating any product that is not a palindrome or is not larger than the largest previously-found palindrome, and 2) using the lower number of any pair that results in a palindrome to set a lower bound for the factors of any potential future palindromes. This rapidly shrinks the search space.