Complete solution to problem 0007
Finished the problem with two functions. One of the functions checks if a given number 'i' is prime by checking it against all the previously found primes in the array primeArray (which starts with 2 already in it). The other function iterates through all odd numbers until primeArray.length is equal to targetNum, which is the n for the nth prime number you're looking for; while looping through the odd numbers, it calls the first function to check if each number is prime, and if so, adds it to the end of primeArray. The function returns the last element of primeArray when this is complete.